fix bugs

+++ +nn/+generator/xavier.m
@@ -1,4 +1,4 @@
 function res = xavier(dimensionVector, ~)
-    c = prod(dimensionVector)/dimensionVector(4);
-    res = (rand(dimensionVector, 'single')-0.5)*2*sqrt(3/c);
+    c = single(prod(dimensionVector)/dimensionVector(4));
+    res = (rand(dimensionVector, 'single')-single(0.5))*single(2)*sqrt(single(3)/c);
 end
